Switching sequencer.S to integer + Applying optimisation to footsteps_mpc

+++ ContactSequencer.py
@@ -22,7 +22,7 @@ class ContactSequencer:
         self.phases = np.array([[0, 0.5, 0.5, 0]])
 
         # Contact sequence
-        self.S = np.zeros((int(self.T_gait/self.dt), 4))
+        self.S = np.zeros((np.int(self.T_gait/self.dt), 4), dtype=np.int)
 
         # Create contact sequence
         self.createSequence()
@@ -40,13 +40,13 @@ class ContactSequencer:
         """
 
         t_seq = np.linspace(t, t+self.T_gait-self.dt, int(np.round(self.T_gait/self.dt))).reshape((-1, 1))
-        phases_seq = (np.hstack((t_seq, t_seq, t_seq, t_seq)) - self.phases * self.T_gait) * 2 * np.pi / self.T_gait
+        phases_seq = (np.hstack((t_seq, t_seq, t_seq, t_seq)) - self.phases * np.float(self.T_gait)) * 2 * np.pi / self.T_gait
         self.S = (np.sin(phases_seq) >= 0).astype(int)
 
         # To have a four feet stance phase at the start
         # Likely due to numerical effect we don't have it
         self.S[0, :] = np.ones((1, 4), dtype=np.int)
-        self.S[int(self.S.shape[0]*0.5), :] = np.ones((1, 4), dtype=np.int)
+        self.S[np.int(self.S.shape[0]*0.5), :] = np.ones((1, 4), dtype=np.int)
 
         # For 4 feet stance phase
         # self.S = np.matrix(np.ones(self.S.shape))
